完成第一类for循环流程控制语句解析; 重写核心解析代码，调整命名，整理代码，加快解析速度；修复一些小bug
